+++ src/main/java/subway/presentation/GlobalExceptionHandler.java
@@ -1,0 +1,59 @@
+package subway.presentation;
+
+import java.util.stream.Collectors;
+import org.springframework.context.support.DefaultMessageSourceResolvable;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+import subway.presentation.dto.response.ExceptionResponse;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
+
+    private static final String EXCEPTION_FORMAT = "%s : ";
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ExceptionResponse> handleException(final Exception ex) {
+        logger.error(String.format(EXCEPTION_FORMAT, "Exception"), ex);
+
+        return ResponseEntity.internalServerError().body(new ExceptionResponse("예상치 못한 예외가 발생했습니다."));
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<ExceptionResponse> handleIllegalArgumentException(final IllegalArgumentException ex) {
+        logger.info(String.format(EXCEPTION_FORMAT, "IllegalArgumentException"), ex);
+
+        return ResponseEntity.badRequest().body(new ExceptionResponse(ex.getMessage()));
+    }
+
+    @Override
+    protected ResponseEntity<Object> handleMethodArgumentNotValid(final MethodArgumentNotValidException ex,
+            final HttpHeaders headers, final HttpStatus status, final WebRequest request) {
+        logger.info(String.format(EXCEPTION_FORMAT, "MethodArgumentNotValidException"), ex);
+
+        final String fieldErrorMessages = ex.getBindingResult().getFieldErrors()
+                .stream()
+                .map(DefaultMessageSourceResolvable::getDefaultMessage)
+                .collect(Collectors.joining(System.lineSeparator()));
+
+        return ResponseEntity.badRequest().body(new ExceptionResponse(fieldErrorMessages));
+    }
+
+    @Override
+    protected ResponseEntity<Object> handleExceptionInternal(
+            final Exception ex,
+            final Object body,
+            final HttpHeaders headers,
+            final HttpStatus status,
+            final WebRequest request
+    ) {
+        logger.warn(String.format(EXCEPTION_FORMAT, ex.getClass().getSimpleName()), ex);
+
+        return super.handleExceptionInternal(ex, body, headers, status, request);
+    }
+}
